Improve integer type checking in `Utility::GetBuffer()`
The inconsistent types were also observed on 32-bit Linux.

In ROOT, this was hotfixed by adding a check for 32-bit Linux in the
preprecessor check, but this required including `RConfig.h`. I don't
think this is desirable.

A better and more direct check for the underlying problem case is to
check if `long int` and `int` have the same size.

This is part of the effort to reduce the differences between PyROOT and
upstream CPyCppyy.

Related commits:
* wlav/CPyCppyy@8a72866
* root-project@02251b5