Do not use file globbing in the build system
This makes the build configuration more robust to added and removed
files.

See also the analogous PR in ROOT, as well as the corresponding Jira
ticket:
  * https://github.com/root-project/root/pull/3259
  * https://its.cern.ch/jira/browse/ROOT-5998

Like this, we also don't need to "comment out" whole source files with
preprocessor macros, like the `TPyClassGenerator`. This also helps ROOT,
which requires these sources, because then the sources can be used
as-is.