Improve heuristic for Clone methods
The heuristic checks if the token "Clone" is present in the method name being
called and gives Python the ownership of the returned object. This is valid
behaviour for methods such as `TObject::Clone` where the ownership of the
returned value is given to the caller. But the `CPPOverload::Set` method gets as
the method name being called the actual function name concatenated with the full
template argument list. For example, given the function

```
template <typename T>
void foo();
```

When it passes through this method its name will be `foo<T>`. This behaviour has
an undesired side-effect. When `T` is a type with the token `Clone` in its name,
then the heuristic will trigger and Python will take ownership of the return
value. This is the root cause of the problem seen at
https://github.com/root-project/root/issues/16725, where `T` is `TClonesArray`.

This commit improves on the heuristic by stripping the template argument list
from the method name before checking for the presence of the token.

Co-authored-by: Jonas Rembser <jonas.rembser@cern.ch>