[CPyCppyy] Remove local memory policy feature
Steering the memory policy locally at the level of methods and functions
is redundant, because one can always define pythonizations that take
care of that.

Removing this feature allows us to add new policy flags without
increasing the net complexity of the library.

With this feature removed, one single bool flag to indicate the
heuristic memory policy is enough, because there is no need to use 4
bits to encode the combination of local and global policies anymore.