Split up `CPyCppyyModule.cxx` up into two files
In ROOT, we're actually using CPyCppyy in two ways:

  1. As a library, that ROOT's TPython library is linked against

  2. As the CPython extension that is imported from Python

The problem with this is if the Python site packages directory is not
the same as the ROOT library directory, we needed to install the library
in two places so both Python and TPython can find it. This was solved
with system links.

The problem with system links is that they are not allowed in the ROOT
Python wheels.

To fix this problem, we split up CPyCppyy into a "normal" library, and a
CPython extension that is linked against this library. Then, we put the
relative path to the library part into the `rpath` of the CPython
extension, and all works fine.

This requires however to have some source file for the CPython extension
part, which was solved in a minimal fashion by having just one function
that initializes the module by calling the library. In other words: the
`CPyCppyyModule.cxx` was split into a library part, and another
translation unit that contains the `extern "C" PyObject* PyInit_libcppyy()`.

It would be great if this can be done in upstream CPyCppyy as well, to
keep the code in sync. For upstream CPyCppyy, this should be
non-controversial, because the `CMakeLists.txt` is not changed. So the
resulting shared library is still the same, the split into different
translation units is just different.